refactor(permalink): extract permalink parsing into its own module
To be used in pagination
refactor(permalink): move tests to permalink.rs
