+++ src/document.rs
@@ -13,6 +13,7 @@ use rss;
 
 use cobalt_model;
 use cobalt_model::files;
+use cobalt_model::permalink;
 use cobalt_model::slug;
 use error::*;
 
@@ -92,54 +93,6 @@ fn permalink_attributes(front: &cobalt_model::Frontmatter, dest_file: &Path) -> 
     attributes.insert("data".to_owned(), Value::Object(front.data.clone()));
 
     attributes
-}
-
-fn explode_permalink<S: AsRef<str>>(permalink: S, attributes: &liquid::Object) -> Result<String> {
-    explode_permalink_string(permalink.as_ref(), attributes)
-}
-
-fn explode_permalink_string(permalink: &str, attributes: &liquid::Object) -> Result<String> {
-    lazy_static! {
-        static ref PERMALINK_PARSER: liquid::Parser = liquid::Parser::new();
-    }
-    let p = PERMALINK_PARSER.parse(permalink)?;
-    let mut p = p.render(attributes)?;
-
-    // Handle the user doing windows-style
-    p = p.replace("\\", "/");
-
-    // Handle cases where substutions were blank
-    p = p.replace("//", "/");
-
-    if p.starts_with('/') {
-        p.remove(0);
-    }
-
-    Ok(p)
-}
-
-fn format_url_as_file<S: AsRef<str>>(permalink: S) -> PathBuf {
-    format_url_as_file_str(permalink.as_ref())
-}
-
-fn format_url_as_file_str(permalink: &str) -> PathBuf {
-    let mut path = Path::new(&permalink);
-
-    // remove the root prefix (leading slash on unix systems)
-    if path.has_root() {
-        let mut components = path.components();
-        components.next();
-        path = components.as_path();
-    }
-
-    let mut path_buf = path.to_path_buf();
-
-    // explode the url if no extension was specified
-    if path_buf.extension().is_none() {
-        path_buf.push("index.html")
-    }
-
-    path_buf
 }
 
 fn document_attributes(
@@ -234,9 +187,9 @@ impl Document {
 
         let (file_path, url_path) = {
             let perma_attributes = permalink_attributes(&front, rel_path);
-            let url_path = explode_permalink(&front.permalink, &perma_attributes)
+            let url_path = permalink::explode_permalink(&front.permalink, &perma_attributes)
                 .chain_err(|| format!("Failed to create permalink `{}`", front.permalink))?;
-            let file_path = format_url_as_file(&url_path);
+            let file_path = permalink::format_url_as_file(&url_path);
             (file_path, url_path)
         };
 
@@ -454,44 +407,5 @@ mod test {
         let input = Path::new("./hello/world/file.liquid");
         let actual = format_path_variable(input);
         assert_eq!(actual, "hello/world");
-    }
-
-    #[test]
-    fn explode_permalink_relative() {
-        let attributes = liquid::Object::new();
-        let actual = explode_permalink("relative/path", &attributes).unwrap();
-        assert_eq!(actual, "relative/path");
-    }
-
-    #[test]
-    fn explode_permalink_absolute() {
-        let attributes = liquid::Object::new();
-        let actual = explode_permalink("/abs/path", &attributes).unwrap();
-        assert_eq!(actual, "abs/path");
-    }
-
-    #[test]
-    fn explode_permalink_blank_substitution() {
-        let attributes = liquid::Object::new();
-        let actual = explode_permalink("//path/middle//end", &attributes).unwrap();
-        assert_eq!(actual, "path/middle/end");
-    }
-
-    #[test]
-    fn format_url_as_file_absolute() {
-        let actual = format_url_as_file("/hello/world.html");
-        assert_eq!(actual, Path::new("hello/world.html"));
-    }
-
-    #[test]
-    fn format_url_as_file_no_explode() {
-        let actual = format_url_as_file("/hello/world.custom");
-        assert_eq!(actual, Path::new("hello/world.custom"));
-    }
-
-    #[test]
-    fn format_url_as_file_explode() {
-        let actual = format_url_as_file("/hello/world");
-        assert_eq!(actual, Path::new("hello/world/index.html"));
     }
 }
